Move service check tests to separate class
