+++ src/test/java/org/datadog/jmxfetch/TestApp.java
@@ -1,11 +1,6 @@
 package org.datadog.jmxfetch;
 
-import com.beust.jcommander.JCommander;
-
-import org.apache.log4j.Level;
-import org.datadog.jmxfetch.reporter.Reporter;
 import org.datadog.jmxfetch.reporter.ConsoleReporter;
-import org.datadog.jmxfetch.util.CustomLogger;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
@@ -26,23 +21,8 @@ public class TestApp {
 
     @BeforeClass
     public static void init() {
-        CustomLogger.setup(Level.toLevel("ALL"), "/tmp/jmxfetch_test.log");
+        CommonTestSetup.setupLogger();
     }
-
-    public static App initApp(String yamlFileName, AppConfig appConfig){
-        // We do a first collection
-        // We initialize the main app that will collect these metrics using JMX
-        String confdDirectory = Thread.currentThread().getContextClassLoader().getResource(yamlFileName).getPath();
-        confdDirectory = new String(confdDirectory.substring(0, confdDirectory.length() - yamlFileName.length()));
-        String[] params = {"--reporter", "console", "-c", yamlFileName, "--conf_directory", confdDirectory, "collect"};
-        new JCommander(appConfig, params);
-
-        App app = new App(appConfig);
-        app.init(false);
-
-        return app;
-    }
-
 
     @Test
     public void testBeanTags() throws Exception {
@@ -54,7 +34,7 @@ public class TestApp {
 
         // Initializing application
         AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx_bean_tags.yaml", appConfig);
+        App app = CommonTestSetup.initApp("jmx_bean_tags.yaml", appConfig);
 
         // Collecting metrics
         app.doIteration();
@@ -97,7 +77,7 @@ public class TestApp {
 
         // Initializing application
         AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx_domain_include.yaml", appConfig);
+        App app = CommonTestSetup.initApp("jmx_domain_include.yaml", appConfig);
 
         // Collecting metrics
         app.doIteration();
@@ -121,7 +101,7 @@ public class TestApp {
 
         // Initializing application
         AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx_domain_exclude.yaml", appConfig);
+        App app = CommonTestSetup.initApp("jmx_domain_exclude.yaml", appConfig);
 
         // Collecting metrics
         app.doIteration();
@@ -149,7 +129,7 @@ public class TestApp {
 
         // Initializing application
         AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx_domain_regex.yaml", appConfig);
+        App app = CommonTestSetup.initApp("jmx_domain_regex.yaml", appConfig);
 
         // Collecting metrics
         app.doIteration();
@@ -173,7 +153,7 @@ public class TestApp {
 
         // Initializing application
         AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx_list_params_include.yaml", appConfig);
+        App app = CommonTestSetup.initApp("jmx_list_params_include.yaml", appConfig);
 
         // Collecting metrics
         app.doIteration();
@@ -196,7 +176,7 @@ public class TestApp {
 
         // Initializing application
         AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx_list_params_include.yaml", appConfig);
+        App app = CommonTestSetup.initApp("jmx_list_params_include.yaml", appConfig);
 
         // Collecting metrics
         app.doIteration();
@@ -218,7 +198,7 @@ public class TestApp {
 
         // Initializing application
         AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx_list_params_exclude.yaml", appConfig);
+        App app = CommonTestSetup.initApp("jmx_list_params_exclude.yaml", appConfig);
 
         // Collecting metrics
         app.doIteration();
@@ -240,7 +220,7 @@ public class TestApp {
 
         // Initializing application
         AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx_list_beans_include.yaml", appConfig);
+        App app = CommonTestSetup.initApp("jmx_list_beans_include.yaml", appConfig);
 
         // Collecting metrics
         app.doIteration();
@@ -266,7 +246,7 @@ public class TestApp {
 
         // Initializing application
         AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx_list_beans_regex_include.yaml", appConfig);
+        App app = CommonTestSetup.initApp("jmx_list_beans_regex_include.yaml", appConfig);
 
         // Collecting metrics
         app.doIteration();
@@ -294,7 +274,7 @@ public class TestApp {
 
         // Initializing application
         AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx_list_beans_regex_exclude.yaml", appConfig);
+        App app = CommonTestSetup.initApp("jmx_list_beans_regex_exclude.yaml", appConfig);
 
         // Collecting metrics
         app.doIteration();
@@ -322,7 +302,7 @@ public class TestApp {
 
         // Initializing application
         AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx_list_beans_exclude.yaml", appConfig);
+        App app = CommonTestSetup.initApp("jmx_list_beans_exclude.yaml", appConfig);
 
         // Collecting metrics
         app.doIteration();
@@ -336,185 +316,6 @@ public class TestApp {
         mbs.unregisterMBean(excludeMeToo);
     }
 
-
-    @Test
-    public void testServiceCheckOK() throws Exception {
-        // We expose a few metrics through JMX
-        MBeanServer mbs = ManagementFactory.getPlatformMBeanServer();
-        ObjectName objectName = new ObjectName("org.datadog.jmxfetch.test:type=ServiceCheckTest");
-        SimpleTestJavaApp testApp = new SimpleTestJavaApp();
-        mbs.registerMBean(testApp, objectName);
-
-        // We do a first collection
-        AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx.yaml", appConfig);
-
-        app.doIteration();
-        ConsoleReporter reporter = ((ConsoleReporter) appConfig.getReporter());
-        // Test that an OK service check status is sent
-        LinkedList<HashMap<String, Object>> serviceChecks = reporter.getServiceChecks();
-
-        assertEquals(1, serviceChecks.size());
-        HashMap<String, Object> sc = serviceChecks.getFirst();
-        assertNotNull(sc.get("name"));
-        assertNotNull(sc.get("status"));
-        assertNull(sc.get("message"));
-        assertNotNull(sc.get("tags"));
-
-        String scName = (String) (sc.get("name"));
-        String scStatus = (String) (sc.get("status"));
-        String[] scTags = (String[]) (sc.get("tags"));
-
-        assertEquals(Reporter.formatServiceCheckPrefix("jmx"), scName);
-        assertEquals(Status.STATUS_OK, scStatus);
-        assertEquals(scTags.length, 1);
-        assertTrue(Arrays.asList(scTags).contains("instance:jmx_test_instance"));
-        mbs.unregisterMBean(objectName);
-    }
-
-    @Test
-    public void testServiceCheckWarning() throws Exception {
-        MBeanServer mbs = ManagementFactory.getPlatformMBeanServer();
-        ObjectName objectName = new ObjectName("org.datadog.jmxfetch.test:type=ServiceCheckTest");
-
-        //  Test application
-        SimpleTestJavaApp testApp = new SimpleTestJavaApp();
-        // Populate it with a lot of metrics (>350) !
-        testApp.populateHashMap(400);
-
-        // Exposing a few metrics through JMX
-        mbs.registerMBean(testApp, objectName);
-
-        AppConfig appConfig = new AppConfig();
-        App app = initApp("too_many_metrics.yaml", appConfig);
-
-        // JMX configuration should return > 350 metrics
-        app.doIteration();
-        ConsoleReporter reporter = ((ConsoleReporter) appConfig.getReporter());
-
-        // Test that an WARNING service check status is sent
-        LinkedList<HashMap<String, Object>> serviceChecks = reporter.getServiceChecks();
-        LinkedList<HashMap<String, Object>> metrics = reporter.getMetrics();
-        assertTrue(metrics.size() >= 350 );
-
-        assertEquals(1, serviceChecks.size());
-        HashMap<String, Object> sc = serviceChecks.getFirst();
-        assertNotNull(sc.get("name"));
-        assertNotNull(sc.get("status"));
-
-        // Message should not be null anymore and reports a high number of metrics warning
-        assertNotNull(sc.get("message"));
-        assertNotNull(sc.get("tags"));
-
-        String scName = (String) (sc.get("name"));
-        String scStatus = (String) (sc.get("status"));
-        String[] scTags = (String[]) (sc.get("tags"));
-
-        assertEquals(Reporter.formatServiceCheckPrefix("too_many_metrics"), scName);
-        // We should have a warning status
-        assertEquals(Status.STATUS_WARNING, scStatus);
-        assertEquals(scTags.length, 1);
-        assertTrue(Arrays.asList(scTags).contains("instance:jmx_test_instance"));
-        mbs.unregisterMBean(objectName);
-    }
-
-    @Test
-    public void testServiceCheckCRITICAL() throws Exception {
-        // Test that a non-running service sends a critical service check
-        MBeanServer mbs = ManagementFactory.getPlatformMBeanServer();
-        ObjectName objectName = new ObjectName("org.datadog.jmxfetch.test_non_running:type=ServiceCheckTest2");
-        SimpleTestJavaApp testApp = new SimpleTestJavaApp();
-        mbs.registerMBean(testApp, objectName);
-
-        AppConfig appConfig = new AppConfig();
-
-        App app = initApp("non_running_process.yaml", appConfig);
-        ConsoleReporter reporter = ((ConsoleReporter) appConfig.getReporter());
-
-        // Test that a CRITICAL service check status is sent on initialization
-        LinkedList<HashMap<String, Object>> serviceChecks = reporter.getServiceChecks();
-        assertEquals(1, serviceChecks.size());
-
-        HashMap<String, Object> sc = serviceChecks.getFirst();
-        assertNotNull(sc.get("name"));
-        assertNotNull(sc.get("status"));
-        assertNotNull(sc.get("message"));
-        assertNotNull(sc.get("tags"));
-
-        String scName = (String) (sc.get("name"));
-        String scStatus = (String) (sc.get("status"));
-        String scMessage = (String) (sc.get("message"));
-        String[] scTags = (String[]) (sc.get("tags"));
-
-        assertEquals(Reporter.formatServiceCheckPrefix("non_running_process"), scName);
-        assertEquals(Status.STATUS_ERROR, scStatus);
-        assertEquals("Cannot connect to instance process_regex: .*non_running_process_test.* Cannot find JVM matching regex: .*non_running_process_test.*", scMessage);
-        assertEquals(scTags.length, 1);
-        assertTrue(Arrays.asList(scTags).contains("instance:jmx_test_instance"));
-
-
-        // Test that a CRITICAL service check status is sent on iteration
-        app.doIteration();
-
-        serviceChecks = reporter.getServiceChecks();
-        assertEquals(1, serviceChecks.size());
-
-        sc = serviceChecks.getFirst();
-        assertNotNull(sc.get("name"));
-        assertNotNull(sc.get("status"));
-        assertNotNull(sc.get("message"));
-        assertNotNull(sc.get("tags"));
-
-        scName = (String) (sc.get("name"));
-        scStatus = (String) (sc.get("status"));
-        scMessage = (String) (sc.get("message"));
-        scTags = (String[]) (sc.get("tags"));
-
-        assertEquals(Reporter.formatServiceCheckPrefix("non_running_process"), scName);
-        assertEquals(Status.STATUS_ERROR, scStatus);
-        assertEquals("Cannot connect to instance process_regex: .*non_running_process_test.*. Is a JMX Server running at this address?", scMessage);
-        assertEquals(scTags.length, 1);
-        assertTrue(Arrays.asList(scTags).contains("instance:jmx_test_instance"));
-
-        mbs.unregisterMBean(objectName);
-    }
-
-    @Test
-    public void testServiceCheckCounter() throws Exception {
-        AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx.yaml", appConfig);
-        Reporter repo = appConfig.getReporter();
-
-        // Let's check that the counter is null
-        assertEquals(0, repo.getServiceCheckCount("jmx"));
-
-        // Let's put a service check in the pipeline (we cannot call doIteration()
-        // here unfortunately because it would call reportStatus which will flush
-        // the count to the jmx_status.yaml file and reset the counter.
-        repo.sendServiceCheck("jmx", Status.STATUS_OK, "This is a test", null);
-
-        // Let's check that the counter has been updated
-        assertEquals(1, repo.getServiceCheckCount("jmx"));
-
-        // Let's check that each service check counter is reset after each app
-        // app iteration
-        app.doIteration();
-        assertEquals(0, repo.getServiceCheckCount("jmx"));
-    }
-
-    @Test
-    public void testPrefixFormatter() throws Exception {
-        // Let's get a list of Strings to test (add real versionned check names
-        // here when you add  new versionned check)
-        String[][] data = {
-            {"activemq_58.foo.bar12", "activemq.foo.bar12"},
-            {"test_package-X86_64-VER1:0.weird.metric_name", "testpackage.weird.metric_name" }
-        };
-
-        // Let's test them all
-        for(int i=0; i<data.length; ++i)
-            assertEquals(data[i][1], Reporter.formatServiceCheckPrefix(data[i][0]));
-    }
 
     @Test
     public void testExitWatcher() throws Exception {
@@ -541,7 +342,7 @@ public class TestApp {
 
         // We do a first collection
         AppConfig appConfig = new AppConfig();
-        App app = initApp("jmx.yaml", appConfig);
+        App app = CommonTestSetup.initApp("jmx.yaml", appConfig);
 
         app.doIteration();
         ConsoleReporter reporter = ((ConsoleReporter) appConfig.getReporter());
